python/eval-runner/pyproject.toml
[project]
name = "eval-runner"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.28.0",
  "pydantic>=2.11.0",
  "python-dotenv>=1.0.0",
  "ragas>=0.2.0",
  "datasets>=3.0.0",
  "typer>=0.15.0",
  "psycopg2-binary>=2.9.0",
  "langchain-ollama>=0.3.0",
  "numpy>=1.26.0",
  "redis>=5.0.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]

[project.scripts]
eval-runner          = "eval_runner.ragas_batch:main"
realtime-eval        = "eval_runner.realtime_eval:main"
query-runner         = "eval_runner.query_runner:main"
prepare-eval-data    = "eval_runner.prepare_eval_data:main"
ingest-training-docs = "eval_runner.ingest_training_docs:main"
bulk-query-runner    = "eval_runner.bulk_query_runner:main"
citizen-query-gen    = "eval_runner.citizen_query_gen:main"
cluster-questions    = "eval_runner.cluster_questions:main"
export-db-eval-data  = "eval_runner.export_db_eval_data:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/eval_runner"]
